% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_functions.R
\name{add_path}
\alias{add_path}
\title{Create PATHs to add to MATLAB PATHs}
\usage{
add_path(path)
}
\arguments{
\item{path}{path to add}
}
\value{
A character vector
}
\description{
Create PATHs to add to MATLAB PATHs
}
\examples{
add_path("~/")
}
