% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BEK_ica_overdecompose}
\alias{BEK_ica_overdecompose}
\title{Decomposition of a transcriptome data}
\format{a list containg \describe{
 \item{A}{A ICA matrix (sample scores)}
 \item{S}{S ICA matrix (gene scores)}
 \item{names}{gene names}
 \item{samples}{sample names}
 \item{counts}{raw counts (non centered)}
 \item{log.counts}{log2 counts (non centered)}
 }}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE23720}
}
\usage{
BEK_ica_overdecompose
}
\description{
A dataset overdecomposed (into 100 components). Data were downloaded from GEO,
then \code{\link{run_fastica}} using MATLAB algorithm with stabilisation as
applied.
}
\details{
Source:
Bekhouche I, Finetti P, Adelaïde J, Ferrari A et al. High-resolution
comparative genomic hybridization of inflammatory breast cancer and
identification of candidate genes. PLoS One 2011 Feb 9;6(2):e16950. PMID:
21339811
}
\keyword{datasets}
