% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Biton.list}
\alias{Biton.list}
\title{Array of all Metagenes}
\format{list of 11 elements}
\source{
\url{http://www.cell.com/cell-reports/abstract/S2211-1247(14)00904-8}
}
\usage{
Biton.list
}
\description{
list of metagenes
}
\keyword{datasets}
