% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CAF.list}
\alias{CAF.list}
\title{CAF signatures from brest cancer}
\format{list of 4 data.frames}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE37614}
}
\usage{
CAF.list
}
\description{
Signatures of 4 subtypes of CAF
}
\details{
Tchou J, Kossenkov AV, Chang L, Satija C et al. Human breast cancer associated fibroblasts exhibit subtype specific gene expression profiles. BMC Med Genomics 2012 Sep 6;5:39. PMID: 22954256
}
\keyword{datasets}
