% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_cell_types.R
\name{cell_voting_immgen}
\alias{cell_voting_immgen}
\title{Attribute cell type to a component}
\usage{
cell_voting_immgen(enrich, n = 10)
}
\arguments{
\item{enrich}{enrichment reults from \code{\link{gene_enrichment_test}}}

\item{n}{look into \code{n} top results}
}
\value{
it returns list of data frames for each result of enrichment list
  from  \code{\link{gene_enrichment_test}}
}
\description{
Works only with immgen signatures
}
\examples{
set.seed(123)
res_run_ica <- run_fastica (
 Example_ds,
 optimal = FALSE,
 n.comp = 20,
 with.names = TRUE
)
corr <- correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)

assign <- assign_metagenes(corr$r)

immune_c<- identify_immune_ic(corr$r[,"M8_IMMUNE"], assign[, "IC"], threshold = 0.1)

enrichment <- gene_enrichment_test(
res_run_ica$S,
res_run_ica$names,
names(immune_c),
alternative = "greater",
p.adjust.method = "BH",
n = 50,
n.consider = 100,
p.value.threshold = 0.05
)

cell_voting_immgen(enrichment$enrichment)
}
