% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_cell_types.R
\name{cell_voting_immgene}
\alias{cell_voting_immgene}
\title{Attribute cell type to a component}
\usage{
cell_voting_immgene(enrich, n = 10)
}
\arguments{
\item{enrich}{enrichment reults from \code{\link{get_enrichment}}}

\item{n}{look into \code{n} top results}
}
\value{
it returns list of data frames for each result of enrichment list
  from  \code{\link{get_enrichment}}
}
\description{
Attribute cell type to a component
}
\examples{
\dontrun{
CIT.cell.types.immgen <- identify_cell_types(S = res.pipeline.BRCACIT$ica$S,res.pipeline.BRCACIT$ica$names, immune.ics=res.pipeline.BRCACIT$immune.ics)
cell_voting_immgene (CIT.cell.types.immgen$enrich, 10)
}
}
