% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconica.R
\docType{package}
\name{deconica}
\alias{deconica}
\alias{deconica-package}
\title{deconICA: Deconvolution of transcriptome through Immune Component Analysis}
\description{
deconICA is a package to perform unsupervised deconvolution of complex
mixtures, it contains functions implementing the pipeline of data
interpretation
}
\details{
See the README on
\href{link}{CRAN}
or \href{https://github.com/}{GitHub}
}
\section{deconICA functions}{
NA
}

