% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_test_samples.R
\name{dist_test_samples}
\alias{dist_test_samples}
\title{Title}
\usage{
dist_test_samples(assign, A, sample.names, quant, normalized.count.df,
  test.type, wide = TRUE, isLog = NULL)
}
\arguments{
\item{assign}{result of \code{\link{assign_metagenes}}}

\item{A}{result of \code{\link{run_fastica}} the \code{A} matrix}

\item{sample.names}{names of samples, should correspond to number of columns
of \code{A}}

\item{quant}{quantiles to use, in form of \code{c(x, y)}}

\item{normalized.count.df}{matrix of counts used for \code{\link{run_fastica}} if
data in log provided, it will be un-log to counts, provide \code{isLog},
best to use \code{raw.counts} or \code{log.counts} from  \code{\link{run_fastica}} to avoid problem
with duplicated gene names}

\item{test.type}{test of distributions to perform}

\item{wide}{if \code{TRUE} returns n by n.ics matrix, if \code{FALSE} returns
long format convininient for \code{ggplot}}

\item{isLog}{by default \code{NULL}, if \code{normalized.count.df} is not
conts but log, provide the base of log, for natural logarithm use
\code{exp(1)}}
}
\value{
returns a matrix (in long or wide) format
}
\description{
Title
}
\examples{


}
