% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_functions.R
\name{doICA}
\alias{doICA}
\title{Call \code{doICA} matlab function}
\usage{
doICA(df.scaled.t, names, samples, path_global = getwd(), n, name = FALSE,
  export.corr = FALSE, corr_folder = "CORRELATION", matlbpth = NULL,
  fasticapth = paste0(path.package("deconica", quiet = TRUE), "/fastica++"))
}
\arguments{
\item{df.scaled.t}{scaled numerical data matrix}

\item{names}{gene names, no duplicates}

\item{samples}{sample names}

\item{path_global}{path where files will be saved}

\item{n}{number of components}

\item{name}{\code{FALSE} by default, name of dataset is used, you can put
your name}

\item{export.corr}{\code{FALSE} by default, if you want to use a java
correlation function later or select \code{TRUE}}

\item{corr_folder}{\code{"CORRELATION"} by default, only if you selected
\code{export.corr = TRUE}}

\item{matlbpth}{is found automatically with \code{\link{get_matlab_2}} function,
replace if not functional}

\item{fasticapth}{path to \code{fastica++} repository with MATLAB scripts}
}
\value{
it returns \code{A}, \code{S} matrices of ICA and \code{names} and
\code{samples} for coherence
}
\description{
function used inside \code{\link{run_fastica}} to run \code{fastICA} with
\code{icasso} stabilization. Matlab engine is necessary
}
\examples{
\dontrun{
data(Example_ds)
res.pre <-
 prepare_data_for_ica(Example_ds[, -1], names = Example_ds[, 1])
res.do <- doICA(
  df.scaled.t =  res.pre$df.scaled,
  names = res.pre$names,
  samples = res.pre$samples,
  path_global = getwd(),
  n = 5,
  name = "test",
  export.corr = FALSE
)
}

}
\seealso{
\code{\link[matlabr]{get_matlab}}, \code{\link{run_fastica}},
\code{\link{export_for_ICA}},\code{\link[matlabr]{run_matlab_code}},
\code{\link{import_ICA_res}}, code{\link{export_for_correlation_java}}
}
