% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_functions.R
\name{doICABatch}
\alias{doICABatch}
\title{\code{doBatchICA}}
\usage{
doICABatch(df, vec, path_global = getwd(), names, samples, name = NULL,
  matlbpth = matlabr::get_matlab(),
  fasticapth = paste0(path.package("deconica", quiet = TRUE), "/fastica++"))
}
\arguments{
\item{df}{numerical data matrix}

\item{vec}{vector of values for which ICA should be computed}

\item{path_global}{path were files will be saved, current directory by default}

\item{names}{gene names}

\item{samples}{sample names}

\item{name}{name of the dataset, if not provided, name of R variable}

\item{matlbpth}{path to matlab, found automaticaly with \code{\link{get_matlab}}}

\item{fasticapth}{path to \code{fastica++}}
}
\value{
plots of stability and MSTD if possible
}
\description{
prepares the data (scales and removes duplicates), runs \code{doBatchICA.m} MATLAB script
}
\examples{
\dontrun{
data(Example_ds)
doICABatch(
  Example_ds[, -1],
  seq(2, 4, 1),
  names = Example_ds[, 1],
  samples = colnames(Example_ds[, -1]),
  name = "test",
  fasticapth = paste0(path.package("deconica", quiet = FALSE), "/fastica++")
)
}
}
