% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_functions.R
\name{export_for_correlation_java}
\alias{export_for_correlation_java}
\title{Exports \code{S} ICA matrix in a specific format}
\usage{
export_for_correlation_java(corr_folder = "CORRELATION", names, S, samples, A,
  ncomp, name, path_global_1 = getwd())
}
\arguments{
\item{corr_folder}{export folder name \code{"CORRELATION"} by default}

\item{names}{gene names}

\item{S}{\code{S} ICA matrix}

\item{samples}{sample names}

\item{A}{\code{A} ICA matrix}

\item{ncomp}{number of computed components}

\item{name}{name of the dataset}

\item{path_global_1}{absolute path}
}
\value{
saves on the drive in \code{corr_folder} exported files
}
\description{
needed for an external function in java
}
\examples{
\dontrun{
data(Example_ds)
res.pre <-
 prepare_data_for_ica(Example_ds[, -1], names = Example_ds[, 1])
res.do <- doICA(
  df.scaled.t =  res.pre$df.scaled,
  names = res.pre$names,
  samples = res.pre$samples,
  path_global = getwd(),
  n = 5,
  name = "test",
  export.corr = FALSE
)
export_for_correlation_java(
S = res.do$S,
A = t(res.do$A),
names =  res.do$names,
samples = res.do$samples,
name = "test",
ncomp = 5
)
}
}
\seealso{
\code{\link{run_fastica}}, \code{\link{import_ICA_res}},
\code{\link{doICA}}, \code{\link{export_for_ICA}}
}
