% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_functions.R
\name{export_for_ICA}
\alias{export_for_ICA}
\title{Export files}
\usage{
export_for_ICA(df.scaled.t, names, samples, path_global = getwd(),
  name = FALSE, n = "")
}
\arguments{
\item{df.scaled.t}{scaled numerical matrix}

\item{names}{gene names, vector of chatracter string}

\item{samples}{sample names, vector of chatracter string}

\item{path_global}{path to export files, current directory by default}

\item{name}{name of the dataset}

\item{n}{number of components}
}
\value{
writes files on the drive in indicated location
}
\description{
export files in right format to run fastICA in MATLAB or BiodICA
}
\examples{
\dontrun{
data(Example_ds)
res.pre <-
 prepare_data_for_ica(Example_ds[, -1], names = Example_ds[, 1])
export_for_ICA(res.pre$df.scaled,
res.pre$names,
res.pre$samples,
path_global = getwd(),
name ="test",
n = 5)
}
}
\seealso{
\code{\link{run_fastica}}, \code{\link{import_ICA_res}},
\code{\link{doICA}}, \code{\link{export_for_correlation_java}}
}
