% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_cell_types.R
\name{gene_enrichment_test}
\alias{gene_enrichment_test}
\title{Enrichment analysis}
\usage{
gene_enrichment_test(S, gene.names, immune.ics, gmt = ImmgenHUGO,
  alternative = c("greater", "lower"), p.adjust.method = c("holm",
  "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"), n = 100,
  n.consider = 500, min_module_size = 5, max_module_size = 500,
  p.value.threshold = 0.05, orient.long = TRUE)
}
\arguments{
\item{S}{matrix of components, dim \code{n} corresponding to genes, \code{m}
corresponding to number of components, use oriented matrix}

\item{gene.names}{character vector of gene names, length needs to be equal to
\code{n}}

\item{immune.ics}{vector of charatcter names of components to use for
enrichment test}

\item{gmt}{data.frame obtained from gmt file with a function
\code{\link[ACSNMineR:format_from_gmt]{format_from_gmt}}, by default Immgen
signatures \url{http://Immgen.org}}

\item{alternative}{greater will check for enrichment, less will check for depletion}

\item{p.adjust.method}{correction method}

\item{n}{number of top genes that wlll be used to test signature}

\item{n.consider}{number of genes from the positive end to be considered}

\item{min_module_size}{minimal module size from gmt file to be considered in
enrichment}

\item{max_module_size}{maximum moudule size from gmt file to be considered in
enrichment}

\item{p.value.threshold}{maximal p-value (corrected if correction is enabled)
that will be displayed}

\item{orient.long}{\code{TRUE} by default, in case you applied transformation to
your \code{S} components, select \code{FALSE}.}
}
\value{
returns value if there is an enrichment in provided
 signatures: \describe{ \item{metagenes}{interpreted metagene gene ranking}
 \item{enrichment}{full results of the enrichment analysis sorted by
 corrected p.value} \item{genes.list}{list of genes used for enrichment} }
}
\description{
Computes an enrichment score (fisher exact test) in provided
 signatures for selected components
}
\details{
\code{gene_enrichment_test} runs enrichment of a
 component (or any ranked list) in known (i.e. immune cell types) signatures.
 It was designed to use \code{S} matrix from  \code{\link{run_fastica}}
 \code{\link{fisher.test}} only on components indentified as correlated with
 immune metagene through function \code{identify_immune_ic} and it searches
 in Immgen signatures \url{http://Immgen.org}.
}
\examples{
set.seed(123)
res_run_ica <- run_fastica (
 Example_ds,
 optimal = FALSE,
 n.comp = 20,
 with.names = TRUE
)
corr <- correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)

assign <- assign_metagenes(corr$r)

immune_c<- identify_immune_comp(corr$r[,"M8_IMMUNE"], assign[, "component"], threshold = 0.1)

gene_enrichment_test(
res_run_ica$S,
res_run_ica$names,
names(immune_c),
alternative = "greater",
p.adjust.method = "BH",
n = 50,
n.consider = 100,
)
}
\seealso{
\code{\link{identify_immune_comp}} identifying immune related
 components, \code{\link{run_fastica}} for running Independent Components
 Analysis, and \code{\link[ACSNMineR]{enrichment}} for enrichment in gmt
 files
}
