% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportions.R
\name{generate_basis}
\alias{generate_basis}
\title{Generate basis matrix}
\usage{
generate_basis(df, sel.ic, markers, orient.long = TRUE)
}
\arguments{
\item{df}{output of \code{run_fastica} containing at least \code{S} and
\code{names} elements}

\item{sel.ic}{ICs indentified as specific sources (i.e. immune cells), by
default it takes all ICs of S matrix, can be provided as valid column names
or numeric index}

\item{markers}{list of markers that should be used for basis matrix (i.e.
"gene.list" from \code{generate_markers}), can be also simple vecor or
list of gene names}

\item{orient.long}{TRUE by default, if you modified S matrix and you don't want it
to be oriented select FALSE}
}
\value{
it returns a \code{data.frame} (basis matrix) that can be used for
  regression or visualization purposes
}
\description{
It generates a basis matrix that can be used for regression from list of
weighted markers
}
\examples{

#generate_basis(df, c("IC1", "IC3", "IC5"), markers = my.gene.list )

#visualize
#basis <- generate_basis(df, c("IC1", "IC3", "IC5"), markers = my.gene.list )
#pheatmap::pheatmap(basis)

}
