% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_cell_types.R
\name{get_enrichment}
\alias{get_enrichment}
\title{Run enrichment analysis for metagenes}
\usage{
get_enrichment(S, gene.names, immune.ics, gmt = ImmgenHUGO,
  quantile = 0.995, abs.thr = NULL, threshold = 0.05, ...)
}
\arguments{
\item{S}{ranks of components, dim \code{n} corresponding to genes, \code{m}
corresponding to number of components}

\item{gene.names}{list of gene names, length needs to be equal to \code{n}}

\item{immune.ics}{list of components to use for enrichment test}

\item{gmt}{list of signatures, by default Immgen signatures
\url{http://Immgen.org}}

\item{quantile}{takes numerical value \code{[0,1]}, default \code{= 0.99}, to
idicate quantile of ranked list that should be used of enrichment, if
\code{ = NULL}, \code{abs.thr} needs to be indicated}

\item{abs.thr}{impose threshold for ranked list, specify  \code{quantile =
NULL} if you want to use \code{abs.thr}}

\item{threshold}{maximal p-value (corrected if correction is enabled) that will be displayed}

\item{...}{other parameters you want to pass to
\code{\link[ACSNMineR]{enrichment}}}
}
\value{
Function returns value if there is an enrichment in provided
  signatures: \describe{ \item{metagenes}{interpreted metagene gene ranking}
  \item{enrich}{full results of the enrichment analysis sorted by corrected
  p.value} }
}
\description{
Verifies enrichment in provided signatures for selected
  metagenes
}
\details{
\code{get_enrichment} runs enrichment of an independent component
  (or any ranked list) in known immune cell types signatures. By default it
  is using \code{S} matrix from ICA \code{\link{run_fastica}} and runs
  enrichment using \code{\link[ACSNMineR]{enrichment}} only on components
  indentified as correlated with immune metagene through function
  \code{identify_immune_ic} and it serches in Immgen signatures
  \url{http://Immgen.org}.
}
\examples{

# run enrichment

}
\seealso{
\code{\link{identify_immune_ic}} identificng immune related
  components, \code{\link{run_fastica}} for running Independent Components
  Analysis, and \code{\link[ACSNMineR]{enrichment}} for enrichment in gmt
  files
}
