% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_functions.R
\name{get_matlab_2}
\alias{get_matlab_2}
\title{Find matlab path}
\usage{
get_matlab_2(try_defaults = TRUE, desktop = FALSE, splash = FALSE,
  display = FALSE, wait = TRUE, mpath = NULL)
}
\arguments{
\item{try_defaults}{(logical) If \code{matlab} is not found from
\code{Sys.which}, and \code{matlab.path} not found, then try some
default PATHs for Linux and OS X.}

\item{desktop}{Should desktop be active for MATLAB?}

\item{splash}{Should splash be active for MATLAB?}

\item{display}{Should display be active for MATLAB?}

\item{wait}{Should R wait for the command to finish.  Both
passed to \code{\link{system}} and adds the \code{-wait} flag.}

\item{mpath}{path to matlab if known}
}
\value{
Character of command for matlab
}
\description{
This tries to find matlab's path using a system which
command, and then, if not found, looks at \code{getOption("matlab.path")}.  If not path is found, it fails.
}
\examples{
if (matlabr::have_matlab()) {
get_matlab_2()
}
}
