% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_metagenes.R
\name{get_max_correlations}
\alias{get_max_correlations}
\title{Assign through maximal correlations}
\usage{
get_max_correlations(corr)
}
\arguments{
\item{corr}{list of correlation matrices with correlation coefficients
and p-values, can be obtained from \code{\link{correlate_metagenes}} or
\code{\link[Hmisc:rcorr]{rcorr}}}
}
\value{
\code{data.frame} with matched column names,
Pearson correlation coefficient, p.value
}
\description{
It assigns maximal correlations between set of correlated vectors
}
\examples{
res_run_ica <- run_fastica (
 Example_ds,
 overdecompose = FALSE,
 n.comp = 5,
 with.names = TRUE
)
corr <- correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)

get_max_correlations(corr)

}
\seealso{
\code{\link[Hmisc:rcorr]{rcorr}}, \code{\link{correlate_metagenes}},
\code{\link{assign_metagenes}}
}
