% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportions.R
\name{get_scores}
\alias{get_scores}
\title{Get aboundance scores}
\usage{
get_scores(df, markers.list, summary = "mean", ...)
}
\arguments{
\item{df}{gene matrix with samples in columns and genes in rows with named
rows}

\item{markers.list}{list of genes or list of genes with weights}

\item{summary}{can be any type of mean i.e. \code{mean}, \code{gm_mean}
(geometric mean), \code{harmonic_mean}, \code{weighted.mean}. For wieghted
mean weights are needed along with gene names}

\item{...}{optional parameters for the mean function}
}
\value{
Function returns numerical value for each column (sample) of provided data frame
}
\description{
It calulates aboundance scores through a mean of marker genes
}
\examples{
set.seed(123)
 res_run_ica <- run_fastica (
 Example_ds,
 optimal = FALSE,
 n.comp = 20,
 with.names = TRUE
)
corr <- correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)

assign <- assign_metagenes(corr$r)

immune <- identify_immune_comp(corr$r[,"M8_IMMUNE"], assign[, "component"], threshold = 0.1)
counts.abs <- (2^res_run_ica$log.counts)-1
row.names(counts.abs) <- res_run_ica$names

markers <- generate_markers(df = res_run_ica,n = 10,
                            sel.ic= names(immune),
                            return= "gene.list")
get_scores (counts.abs, markers, summary = "mean", na.rm = TRUE)

markers <- generate_markers(df = res_run_ica,n  = 10,
                           sel.ic= names(immune),
                           return= "genes.ranked")
get_scores (counts.abs, markers, summary = "weighted.mean", na.rm = TRUE)
}
