% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_metagenes.R
\name{identify_immune_comp}
\alias{identify_immune_comp}
\title{Identify components related to immune signal}
\usage{
identify_immune_comp(x, l, threshold = 0.1)
}
\arguments{
\item{x}{the correaltion with immune metagene can be retreived from
\code{\link{correlate_metagenes}} output}

\item{l}{vector of names of assigned components}

\item{threshold}{lower bound for filtering correlation [0,1]}
}
\value{
it returns data frame of IC names and correlations passing the
\code{threshold}
}
\description{
Identify components related to immune signal
}
\examples{

res_run_ica <- run_fastica (
 Example_ds,
 optimal = FALSE,
 n.comp = 20,
 with.names = TRUE
)
corr <- correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)

assign <- assign_metagenes(corr$r)

identify_immune_comp(corr$r[,"M8_IMMUNE"], assign[, "component"], threshold = 0.1)
}
