% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ImmgenHUGO}
\alias{ImmgenHUGO}
\title{Cell type signatures}
\format{a dataframe with the
\describe{
  \item{module}{first column}
  \item{module length}{second column}
  \item{gene names}{third column}
  }}
\source{
\url{http://www.immgen.org}
}
\usage{
ImmgenHUGO
}
\description{
Imported in correct format with \code{\link[ACSNMineR]{format_from_gmt}} and parsed
}
\keyword{datasets}
