% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_functions.R
\name{import_ICA_res}
\alias{import_ICA_res}
\title{Import results of ICA}
\usage{
import_ICA_res(name, ncomp, path_global_1)
}
\arguments{
\item{name}{name of the dataset}

\item{ncomp}{number of components}

\item{path_global_1}{absolute path of the files}
}
\value{
imports \code{A} and \code{S} ICA matrix
}
\description{
imports files run in Matlab or precomputed
}
\examples{
\dontrun{
data(Example_ds)
res.pre <-
 prepare_data_for_ica(Example_ds[, -1], names = Example_ds[, 1])
res.do <- doICA(
  df.scaled.t =  res.pre$df.scaled,
  names = res.pre$names,
  samples = res.pre$samples,
  path_global = getwd(),
  n = 5,
  name = "test",
  export.corr = FALSE
)
import_ICA_res("test_5", 5, paste0(getwd(),"/test_5/"))
}
}
\seealso{
\code{\link{run_fastica}}, \code{\link{export_for_ICA}},
\code{\link{doICA}}, \code{\link{export_for_correlation_java}}
}
