% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_logscale}
\alias{is_logscale}
\title{Verify if data is in log scale}
\usage{
is_logscale(x)
}
\arguments{
\item{x}{\code{data.frame} or \code{matrix}}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Verify if data is in log scale
}
\examples{
M <- matrix(sample(-1:14, 100, replace = TRUE),10,10, byrow = TRUE)
is_logscale(M)
M2 <- 2^M
is_logscale(M2)
}
