% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LM22.list}
\alias{LM22.list}
\title{Array of all Metagenes}
\format{list of 22 \code{data.frames}}
\source{
\url{http://www.cell.com/cell-reports/abstract/S2211-1247(14)00904-8}
}
\usage{
LM22.list
}
\description{
list of 22 immune cell type profiles
}
\keyword{datasets}
