% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{lolypop_plot_corr}
\alias{lolypop_plot_corr}
\title{Lolypop plot for correlations}
\usage{
lolypop_plot_corr(r, col, head.size = 10, head.color = "value",
  digits = 2, head.text.size = 3.5, head.text.color = "white",
  vertical = TRUE)
}
\arguments{
\item{r}{correlation matrix \code{r} matrix of output \code{\link{correlate_metagenes}}}

\item{col}{select column either index or column name}

\item{head.size}{size of the point of correlation}

\item{head.color}{by default colored by correlation values, if you want one color provide color name}

\item{digits}{parameter of \code{\link[base]{round}} for the correlation showed on the plot. integer
indicating the number of decimal places (round) or significant digits (signif) to be used.}

\item{head.text.size}{size of the correlation text font}

\item{head.text.color}{color of the correlation text font}

\item{vertical}{\code{TRUE} for vertical plot,  \code{FALSE} for horizontal plot}
}
\value{
returns \code{\link[ggplot2]{ggplot}}
}
\description{
Plot correlations between one metagene or known profile and all components in a form of linear plot
which is a variant of a signal plot. Wrapper using ggplot2.
}
\details{
Values are order from highest correlation to lowest correlation. Colors and fonts can be overwritten.
To see all correlations simultaneously choose \code{\link{radar_plot_corr}}
}
\examples{
res_run_ica <- run_fastica (
 Example_ds,
 optimal = FALSE,
 n.comp = 20,
 with.names = TRUE
)
corr <- correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)
#horizontal
lolypop_plot_corr(corr$r,2, vertical =FALSE)
# vertical
lolypop_plot_corr(corr$r,"M8_IMMUNE")
#change colors
lolypop_plot_corr(corr$r,"M8_IMMUNE",head.color = "black" , head.text.color = "green")
#remove title
lolypop_plot_corr(corr$r,"M8_IMMUNE")+ ggplot2::labs(title="",subtitle="")
}
\seealso{
\code{\link[ggplot2]{ggplot}},
  \code{\link[ggplot2]{aes_string}}, \code{\link[ggplot2]{theme_bw}},
  \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{labs}},
  \code{\link[ggplot2]{scale_color_distiller}}, \code{\link[ggplot2]{coord_flip}},
   \code{\link[ggplot2]{geom_segment}}
}
