% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportions.R
\name{make_list}
\alias{make_list}
\title{Make list of weighted markers}
\usage{
make_list(df)
}
\arguments{
\item{df}{\code{data.frame} to be transformed with gene names in the \code{row.names}}
}
\value{
named \code{list} of \code{data.frame}s with gene names in the first column and values in the second column.
}
\description{
Transforms a data frame with multiple columns into a named list of weighted markers
with gene names in the first column and values in the second column.
}
\examples{
X <- as.data.frame(matrix(runif(10000), 50, 10))
row.names(X) <- paste("A",1:nrow(X), sep="")
make_list(X)


}
