% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_test_samples.R
\name{most_variant_IC}
\alias{most_variant_IC}
\title{Compute variance explained by each Independent Component}
\usage{
most_variant_IC(S, A, X, n = 5)
}
\arguments{
\item{S}{result of \code{\link{run_fastica}} the \code{S} matrix}

\item{A}{result of \code{\link{run_fastica}} the \code{A} matrix}

\item{X}{data, either post-PCA data of \code{\link{run_fastica}} \code{X} matrix}

\item{n}{number of top ICs if \code{n} = "all" then fraction of variance explained for
all ICs is returned}
}
\value{
returns a data frame with n top ICs numbers ranked by their fraction of variance explained
}
\description{
Compute variance explained by each Independent Component
}
\examples{
set.seed(123)
res_fastica <- run_fastica (
 Example_ds,
 optimal = FALSE,
 n.comp = 20,
 with.names = TRUE
)
most_variant_IC(res_fastica$S, res_fastica$A, res_fastica$X, n =3)

res <- most_variant_IC(res_fastica$S, res_fastica$A, res_fastica$X, n =5)
barplot(as.matrix(t(res)))

}
