% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_dist_test}
\alias{plot_dist_test}
\title{Plot results of density test}
\usage{
plot_dist_test(df, plot.type = c("line", "density"))
}
\arguments{
\item{df}{\code{data.frame} in long format}

\item{plot.type}{can be either \code{"line"} or \code{"density"}}
}
\value{
returns a line or density plot of p.value or test statistics versus rank or density
}
\description{
Wrapper over \code{\link[ggplot2:ggplot]{ggplot}} ploting either rank or
density versus selected value in \code{\link{dist_test_samples}} (p.value or
test statistics)
}
\examples{
#numerical matrix
set.seed(134)
S <- matrix(stats::rnbinom(10000, mu = 6, size = 10), 500, 80)
dat <- matrix(runif(1600,min =1, max=10 ), 80, 80, byrow = TRUE)
A <- dat / rowSums(dat)
X <- data.frame(S \%*\% A)
res_run_ica <- run_fastica(X, row.center = TRUE, n.comp = 5, optimal = FALSE)

#run the funtion selecting wide = FALSE
res.ttest <- dist_test_samples(A = res_run_ica$A,
sample.names = res_run_ica$samples,
X.counts = res_run_ica$log.counts,
test.type = "t.test",
thr=0.5,
isLog = 2,
return = "p.value",
wide = FALSE)

#plot results
plot_dist_test(res.ttest, plot.type = "density")
plot_dist_test(res.ttest, plot.type = "line")
}
\seealso{
\code{\link[ggplot2]{ggplot}},
  \code{\link[ggplot2]{stat_density}}, \code{\link[ggplot2]{theme_bw}},
  \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{geom_line}}
}
