% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_functions.R
\name{prepare_data_for_ica}
\alias{prepare_data_for_ica}
\title{Formats data for ICA in MATLAB}
\usage{
prepare_data_for_ica(df, names, samples = NULL, isLog = TRUE)
}
\arguments{
\item{df}{numerical data matrix}

\item{names}{gene names character vector}

\item{samples}{if not provided column names will be used}
}
\value{
\item{\code{df.scaled}}{scaled data without duplicates}
\item{\code{names}}{gene names without duplicates}
\item{\code{non.scaled}}{non scaled data without duplicates}
\item{\code{samples}}{sample names}
}
\description{
Formats data for ICA in MATLAB
}
\examples{

data(Example_ds)
prepare_data_for_ica(Example_ds[, -1], names = Example_ds[, 1])
}
\seealso{
\code{\link{run_fastica}}, \code{\link{import_ICA_res}},
\code{\link{doICA}}, \code{\link{doICABatch}}
}
