% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{radar_plot_corr}
\alias{radar_plot_corr}
\title{Radar plot of correlations}
\usage{
radar_plot_corr(df, ax.size = NULL, size.el.txt = 15, point.size = 5)
}
\arguments{
\item{df}{output of function \code{\link{correlate_metagenes}} - correlation
matrix with correlation and p-values}

\item{ax.size}{define size of axis labels, adapts automatically by default}

\item{size.el.txt}{define general size of letters, 15 by default}

\item{point.size}{\code{size} parameter in \code{\link[ggplot2]{geom_point}}}
}
\value{
Radar plots for correlations of each input component with matagene/profile,
Returns a list containing the \code{data.frame} \code{df} used to generate the plot
- long format - and the plot itself \code{p}.
}
\description{
Wrapper using ggplot2 to plot correlations between components and given metagenes or
pure profiles
}
\examples{
res_run_ica <- run_fastica (
 Example_ds,
 optimal = FALSE,
 n.comp = 20,
 with.names = TRUE
)
corr <- correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)

radar_plot_corr(corr)
data <- radar_plot_corr(corr)$df

#change plot
radar_plot_corr(corr)$p +
ggplot2::labs(title="11 Biton et al. metagenes vs my ICA components",
                           subtitle="Pearson correlation coefficients")

radar_plot_corr(corr, point.size = 1)

radar_plot_corr(corr, point.size = 0)$p +
ggplot2::geom_point(size = 8, alpha = 0.4)
}
\seealso{
\code{\link[ggplot2]{ggplot}},
  \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{coord_polar}},
  \code{\link[ggplot2]{theme_bw}}, \code{\link[ggplot2]{facet_wrap}},
  \code{\link[ggplot2]{scale_color_distiller}}, \code{\link[ggplot2]{theme}},
  \code{\link[ggplot2]{element_text}}
}
