% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{radar_plot_markers}
\alias{radar_plot_markers}
\title{Radar plot of correlations}
\usage{
radar_plot_markers(df, ax.size = NULL, size.el.txt = 15)
}
\arguments{
\item{df}{output of function \code{\link{correlate_metagenes}} - correlation
matrix with correlation and p-values}

\item{ax.size}{define general size of letters, 15 by default}

\item{size.el.txt}{define size of axis labels, adapts automatically by default}
}
\value{
It shows radar plots for each input IC and matagene/profile to
  corelate with, it also returns a list containing the data.frame \code{df} used to generate the plot - long
  format - and the plot itself \code{p}.
}
\description{
Wrapper using ggplot2 to plot correlations between ICs and given metagenes or
pure profiles
}
\examples{

#radar_plot_markers(df, alpha =0.2)





}
\seealso{
\code{\link[ggplot2]{ggplot}} components,
  \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{coord_polar}},
  \code{\link[ggplot2]{theme_bw}}, \code{\link[ggplot2]{facet_wrap}},
  \code{\link[ggplot2]{scale_color_distiller}}, \code{\link[ggplot2]{theme}},
  \code{\link[ggplot2]{element_text}}
}
