% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastica.R
\name{run_fastica}
\alias{run_fastica}
\title{Overdecompose dataset with ICA.}
\usage{
run_fastica(X, optimal = TRUE, row.center = TRUE, with.names = FALSE,
  alg.typ = "deflation", method = "C", n.comp = 100, ...)
}
\arguments{
\item{X}{a data matrix with \code{n} rows representing observations and \code{p} columns
representing variables, place gene names in the first column and select
\code{with.names = TRUE} to keep gene names for further analysis}

\item{optimal}{check \code{TRUE} to let select best number of components for
deconvolution, for datasets >120 columns, n.comp will be set to 100, if
<120 then number of components will be selected according to Kaiser Rule
(90 percent of variance explained)}

\item{row.center}{if \code{TRUE} substract row mean from data}

\item{with.names}{if first column of X is row.names please indicate \code{TRUE}, in
case of duplicated names, the transcript with highest variance will be
kept, names need to be HUGO names, if names are not provided at this step,
you can provide them later}

\item{alg.typ}{if \code{alg.typ == "parallel"} the components are extracted
    simultaneously (the default). if \code{alg.typ == "deflation"} the
    components are extracted one at a time.}

\item{method}{if \code{method == "R"} then computations are done
    exclusively in \R (default). The code allows the interested \R user to
    see exactly what the algorithm does.
    if \code{method == "C"} then C code is used to perform most of the
    computations, which makes the algorithm run faster.  During
    compilation the C code is linked to an optimized BLAS library if
    present, otherwise stand-alone BLAS routines are compiled.}

\item{n.comp}{number of components to be extracted}

\item{...}{other possible parameters from \code{\link[fastICA]{fastICA}}}
}
\value{
A list containing the following components as in
  \code{\link[fastICA]{fastICA}} \describe{
  \item{X}{ - pre-processed data matrix (after PCA)}
  \item{K}{ - pre-whitening matrix that projects data onto the first n.comp principal
  components.}
  \item{W}{ - estimated un-mixing matrix (see definition in details)}
  \item{A}{ - estimated mixing matrix}
  \item{S}{ - estimated source matrix}
  \item{names}{ - if  \code{with.names = TRUE} will contain row names list}
  }
}
\description{
This is a wrapper of \code{\link[fastICA]{fastICA}}. It allows compute number
of ICs optimal for over decomposition for immune deconvolution.
}
\examples{
\dontrun{
# numerical matrix
S <- matrix(runif(10000), 5000, 2)
A <- matrix(c(1, 1, -1, 3), 2, 2, byrow = TRUE)
X <- data.frame(S \%*\% A)
run_fastica(X, row.center = TRUE, n.comp = 2, optimal = FALSE)
# matrix with gene names
S <- matrix(runif(10000), 5000, 2)
A <- matrix(c(1, 1, -1, 3), 2, 2, byrow = TRUE)
X <- data.frame(S \%*\% A)
names <- paste("A",1:nrow(X), sep="")
X <- cbind(names,X)
run_fastica(X, row.center = TRUE, n.comp = 2, optimal = FALSE, with.names = TRUE)
}
}
\seealso{
\code{\link[fastICA]{fastICA}}
  \url{https://cran.r-project.org/web/packages/fastICA/index.html}
}
