% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastica.R
\name{run_fastica_import}
\alias{run_fastica_import}
\title{Repriduce process of run_fastica}
\usage{
run_fastica_import(X, overdecompose = TRUE, row.center = TRUE,
  with.names = FALSE, gene.names = NULL, n.comp = 100, isLog = TRUE,
  import = TRUE, path_global = getwd(), name = NULL, ...)
}
\arguments{
\item{X}{a data matrix with \code{n} rows representing observations and
\code{p} columns representing variables, place gene names in the first
column and select \code{with.names = TRUE}}

\item{overdecompose}{check \code{TRUE} to let select best number of components for
deconvolution, for datasets >120 columns, n.comp will be set to 100, if <120
then number of components will be selected according to Kaiser Rule (90
percent of variance explained)}

\item{row.center}{if \code{TRUE} subtract row mean from data}

\item{with.names}{if first column of X is row.names please indicate
\code{TRUE}, in case of duplicated names, the transcript with highest
variance will be kept, names need to be HUGO names, if names are not
provided at this step, you can provide them later}

\item{gene.names}{character vector of row names - gene names}

\item{n.comp}{number of components to be extracted}

\item{isLog}{if data is in log \code{TRUE} if data is in counts \code{FALSE}}

\item{import}{imports data only if TRUE}

\item{path_global}{only if \code{R = FALSE}, the global path where files will
be written, current directory by default}

\item{name}{important for Matlab version, defines the name of your files}

\item{...}{other possible parameters for \code{\link[fastICA]{fastICA}}}
}
\value{
an object as \code{\link{run_fastica}}
}
\description{
Applies preprocessig of \code{\link{run_fastica}} but instead of running ICA
it imports matlab output files. It is handy if you run the matlab idenpendly
or if you lost R session data
}
\examples{
\dontrun{

# numerical matrix
S <- matrix(runif(10000), 10, 2)
A <- matrix(sample(-3:3, 16, replace = TRUE),2,8, byrow = TRUE)
X <- data.frame(S \%*\% A)
#matlab
run_fastica(X, row.center = TRUE, n.comp = 2, overdecompose = FALSE, R = FALSE)
run_fastica_import(X, row.center = TRUE, n.comp = 2, overdecompose = FALSE, import=TRUE)
}


}
