% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_functions.R
\name{run_matlab_code_2}
\alias{run_matlab_code_2}
\title{Runs matlab code}
\usage{
run_matlab_code_2(code, matlbpth = NULL, endlines = TRUE, verbose = TRUE,
  add_clear_all = FALSE, paths_to_add = NULL, ...)
}
\arguments{
\item{code}{Character vector of code.}

\item{matlbpth}{path to matlab engine}

\item{endlines}{Logical of whether the semicolon (;) should be
pasted to each element of the vector.}

\item{verbose}{Print out filename to run}

\item{add_clear_all}{Add \code{clear all;} to the beginning of code}

\item{paths_to_add}{Character vector of PATHs to add to the
script using \code{\link{add_path}}}

\item{...}{Options passed to \code{\link{run_matlab_script}}}
}
\value{
Exit status of matlab code
}
\description{
This function takes in matlab code, where
the last line must end with a ;, and returns the exit
status, slightly modified version of
\code{\link[matlabr]{run_matlab_code}}
}
\examples{
if (matlabr::have_matlab()){
   run_matlab_code_2("disp(version)", matlbpth = "matlbpth")
   run_matlab_code_2("disp(version)", paths_to_add = "~/")
   run_matlab_code_2(c("disp('The version of the matlab is:')", "disp(version)"))
   run_matlab_code_2(c("x = 5", "disp(['The value of x is ', num2str(x)])"))
}
}
