% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_functions.R
\name{run_matlab_script_2}
\alias{run_matlab_script_2}
\title{Run matlab script}
\usage{
run_matlab_script_2(fname, matlbpth = NULL, verbose = TRUE,
  desktop = FALSE, splash = FALSE, display = FALSE, wait = TRUE, ...)
}
\arguments{
\item{fname}{Filename of matlab script (.m file)}

\item{matlbpth}{path to matlab engine}

\item{verbose}{print diagnostic messages}

\item{desktop}{Should desktop be active for MATLAB?}

\item{splash}{Should splash be active for MATLAB?}

\item{display}{Should display be active for MATLAB?}

\item{wait}{Should R wait for the command to finish.  Both
passed to \code{\link{system}} and adds the \code{-wait} flag.}

\item{...}{Options passed to \code{\link{system}}}
}
\value{
Exit status of matlab code
}
\description{
This function runs a matlab script, and
returns exit statuses, slightly modified version of
\code{\link[matlabr]{run_matlab_script}}
}
