% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{stacked_proportions_plot}
\alias{stacked_proportions_plot}
\title{Plot cell proportions}
\usage{
stacked_proportions_plot(dat)
}
\arguments{
\item{dat}{scores data.frame with cell types in lines and samples in columns}
}
\value{
a stacked bar plot based on ggplot2
}
\description{
Plots scores for all samples as a fraction of one in each samples
}
\examples{
#random matrix y
y <- data.frame(matrix(runif(10000), ncol = 100, nrow = 10))
#plot
stacked_proportions_plot(y)
}
