% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TIMER_cellTypes}
\alias{TIMER_cellTypes}
\title{TIMER signatures}
\format{a dataframe with the
\describe{
  \item{module}{first column}
  \item{module length}{second column}
  \item{gene names}{third column}
  }}
\source{
\url{http://cistrome.org/TIMER/}
}
\usage{
TIMER_cellTypes
}
\description{
Signatures of 9 cell types published as part of TIMER tool
}
\details{
Li, Bo, et al. "Comprehensive analyses of tumor immunity: implications for cancer immunotherapy." Genome biology 17.1 (2016): 174.
}
\keyword{datasets}
