% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_metagenes.R
\name{assign_metagenes}
\alias{assign_metagenes}
\title{Assign Independent Component to a metagene through reciprocity}
\usage{
assign_metagenes(r, immune_name = "M8_IMMUNE")
}
\arguments{
\item{r}{the correlation matrix, \code{r} matrix from out of
\code{\link{correlate_metagenes}}}

\item{immune_name}{name of the immune component (if present in \code{r}) to
be excluded from this analysis}
}
\value{
this function returns a data frame with number of IC in the first
column and assigned metagene name in second column
}
\description{
This function assign an idependent component to a metagene through
verification if for an IC the max correlation points to given metagene and if
for this metageneif the max correlation points back the the
}
\examples{

res_run_ica <- run_fastica (
 Example_ds,
 optimal = FALSE,
 n.comp = 5,
 with.names = TRUE
)
corr <- correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)

assign_metagenes(corr$r)


}
