% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_metagenes.R
\name{assign_metagenes}
\alias{assign_metagenes}
\title{Assign components to a metagene through mutual reciprocity}
\usage{
assign_metagenes(r, exclude_name = "M8_IMMUNE")
}
\arguments{
\item{r}{the correlation matrix, \code{r} matrix, can be generated from
\code{\link{correlate_metagenes}} function}

\item{exclude_name}{name of the components (present in \code{r}) to
be excluded from this analysis (for example immune), by default "M8_IMMUNE"
is excluded}
}
\value{
returns a \code{dataf.rame} with component name in the first
column and assigned profile/metagene name in second column
}
\description{
Attributes labels to components under condition of mutual reciprocal correlation
}
\details{
This function assign a component to a metagene/profile through
verification if the component's the maximal correlation points to a given profile and if
for this profile the maximal correlation points back the that component. In mathematical
terms, given correlations between the set of profiles/metagenes \eqn{A = {A_1,...,A_m}} and
\eqn{S} components matrix \eqn{S = {IC1,...,ICN}}, if
\deqn{Si = argmaxi(corr(Aj,S))} and \deqn{A_j = argmax_j(corr(S_i,A))}
}
\examples{

res_run_ica <- run_fastica (
 Example_ds,
 overdecompose = FALSE,
 n.comp = 5,
 with.names = TRUE
)
corr <- correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)

assign_metagenes(corr$r)


}
\seealso{
\code{\link{get_max_correlations}}, \code{\link{correlate_metagenes}}
}
