% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_cell_types.R
\name{cell_voting_immgen}
\alias{cell_voting_immgen}
\title{Attribute cell type to a component}
\usage{
cell_voting_immgen(enrich, n = 10)
}
\arguments{
\item{enrich}{enrichment reults from \code{\link{gene_enrichment_test}}}

\item{n}{look into \code{n} top results}
}
\value{
it returns list of data frames for each result of enrichment list
  from  \code{\link{gene_enrichment_test}}
}
\description{
Works only with immgen signatures
}
\examples{
\dontrun{
CIT.cell.types.immgen <- identify_cell_types(S = res.pipeline.BRCACIT$ica$S,res.pipeline.BRCACIT$ica$names, immune.ics=res.pipeline.BRCACIT$immune.ics)
cell_voting_immgene (CIT.cell.types.immgen$enrich, 10)
}
}
