% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_metagenes.R
\name{correlate_metagenes}
\alias{correlate_metagenes}
\title{Correlates idependent componenets with known ranked lists of metagenes}
\usage{
correlate_metagenes(S, gene.names, metagenes = data.list, threshold = -Inf,
  n.genes.intersect = 30, ...)
}
\arguments{
\item{S}{S matrix of ICA decomposition}

\item{gene.names}{list of gene names, needs to be of the same length as nrow
of S, it is recommended to run \code{\link{run_fastica}} \code{with.names =
TRUE} to assure compatibility}

\item{metagenes}{named list of datasets, each with two columns 1st - gene
names, 2nd - ranks, by default 11 metagenes from Biton et al....}

\item{threshold}{threshold for ICA components (columns of \code{S}) to be
applied before correlation, default set to -Inf (all ranks are kept)}

\item{n.genes.intersect}{minimum of genes that should intersect between IC and
metagene to keep the IC in correlation matrix}

\item{...}{additional params you can pass to \code{\link[Hmisc]{rcorr}}}
}
\value{
provides correlation matrix with correlation cofficient \code{r},
 p.values  \code{P} and number of overlapping genes \code{n}
}
\description{
Independent components obtained with \code{\link{run_fastica}} can be
characterised through correlation with known ranked list (metagenes), by
default this function is using metagenes from Biton et al. (2015) Cell It is
using \code{\link[Hmisc]{rcorr}} function for correlations
}
\examples{

res_run_ica <- run_fastica (
 Example_ds,
 optimal = FALSE,
 n.comp = 5,
 with.names = TRUE
)
correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)


}
\seealso{
\code{\link[Hmisc]{rcorr}} \code{\link{run_fastica}}
}
