% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_ds}
\alias{Example_ds}
\title{Example of dataset}
\format{a dataframe with the
\describe{
  \item{module}{first column}
  \item{module length}{second column}
  \item{gene names}{third column}
 }}
\source{
\url{zezez}
}
\usage{
Example_ds
}
\description{
OVCA
}
\keyword{datasets}
