% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_ds}
\alias{Example_ds}
\title{Example of a cancer dataset}
\format{a dataframe with the \describe{ \item{rows}{21320} \item{columns}{61}
 \item{first column}{is related to GENE names} }}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE23720}
}
\usage{
Example_ds
}
\description{
A a sample 60 randomly selected samples from transcriptome of inflammatory
breast cancer (IBC). Data were centred and in transformed in log2 before sampling
}
\details{
Bekhouche I, Finetti P, Adelaïde J, Ferrari A et al. High-resolution
comparative genomic hybridization of inflammatory breast cancer and
identification of candidate genes. PLoS One 2011 Feb 9;6(2):e16950. PMID:
21339811
}
\keyword{datasets}
