% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportions.R
\name{generate_basis}
\alias{generate_basis}
\title{Generate basis matrix}
\usage{
generate_basis(df, sel.comp, markers, orient.long = TRUE)
}
\arguments{
\item{df}{output of \code{run_fastica} containing at least \code{S} and
\code{names} elements}

\item{sel.comp}{components identified as specific sources (i.e. immune cells), by
default it takes all components of S matrix, can be provided as valid column names
or numeric index}

\item{markers}{list of markers that should be used for basis matrix (i.e.
"gene.list" from \code{generate_markers}), can be also simple vector or
list of gene names}

\item{orient.long}{TRUE by default, if you modified S matrix and you don't want it
to be oriented select FALSE}
}
\value{
it returns a \code{data.frame} (basis matrix) that can be used for
  regression or visualization purposes
}
\description{
It generates a basis matrix that can be used for regression from list of
weighted markers
}
\examples{

set.seed(123)
 res_run_ica <- run_fastica (
 Example_ds,
 overdecompose = FALSE,
 n.comp = 20,
 with.names = TRUE
)
corr <- correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)

assign <- assign_metagenes(corr)

immune <- identify_immune_comp(corr$r[,"M8_IMMUNE"], assign[, "component"], threshold = 0.1)

markers <- generate_markers(df = res_run_ica,n = 10,sel.comp= names(immune), return= "gene.list")
basis <- generate_basis(df = res_run_ica,sel.comp= names(immune),markers= markers )
pheatmap::pheatmap(basis )
}
