% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportions.R
\name{generate_markers}
\alias{generate_markers}
\title{Generate markers from components}
\usage{
generate_markers(df, n = 30, thr = Inf, sel.ic = paste("IC", 1:ncol(df$S),
  sep = ""), return = "gene.list", orient.long = TRUE)
}
\arguments{
\item{df}{output of \code{run_fastica} containing at least \code{S} and \code{names}
elements}

\item{n}{number of top genes considered from each signature, n = 30 by
default}

\item{thr}{max gene expression, if removal of outliers is necessary, Inf (no
threshold) by default.}

\item{sel.ic}{ICs indentified as specific sources (i.e. immune cells), by
default it takes all ICs of S matrix, can be provided as valid column names
or numeric index}

\item{return}{should it return "gene.list" or "metagenes.ranked"}

\item{orient.long}{TRUE by default, if S is oriented change to FALSE}
}
\value{
function returns either list of gene markers "gene.list" for each
  signal (selected IC component) or list of "metagenes.ranked" which are gene
  nemes with ICA weights
}
\description{
It extracts from set of sources (i.e. ICA S matrix) the n top genes(with
weights if needed) to use as marker list or markers with weights for
deconvolution through \code{\link{get_scores}}
}
\examples{
set.seed(123)
 res_run_ica <- run_fastica (
 Example_ds,
 optimal = FALSE,
 n.comp = 20,
 with.names = TRUE
)
corr <- correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)

assign <- assign_metagenes(corr$r)

immune <- identify_immune_ic(corr$r[,"M8_IMMUNE"], assign[, "IC"], threshold = 0.1)

generate_markers(df = res_run_ica,n = 10,sel.ic= names(immune))
generate_markers(df = res_run_ica,n = 10,sel.ic= names(immune), return= "metagenes.ranked")
}
