% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_metagenes.R
\name{get_max_correlations}
\alias{get_max_correlations}
\title{Assign through highest correlations}
\usage{
get_max_correlations(corr)
}
\arguments{
\item{corr}{list of correlation matrices with correlation coefficients
and p-values}
}
\value{
dataframe with column names, correlation coefficient, p.value
}
\description{
It assigns max correlations between set of correlated vectors
}
\examples{
res_run_ica <- run_fastica (
 Example_ds,
 optimal = FALSE,
 n.comp = 5,
 with.names = TRUE
)
corr <- correlate_metagenes(
   S = res_run_ica$S,
   gene.names = res_run_ica$names)

get_max_correlations(corr)

}
