% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportions.R
\name{get_scores}
\alias{get_scores}
\title{Get aboundance scores}
\usage{
get_scores(df, markers.list, summary = "mean", ...)
}
\arguments{
\item{df}{gene matrix with samples in columns and genes in rows with named
rows}

\item{markers.list}{list of genes or list of genes with weights}

\item{summary}{can be any type of mean i.e. \code{mean}, \code{gm_mean}
(geometric mean), \code{harmonic_mean}, \code{weighted.mean}. For wieghted
mean weights are needed along with gene names}

\item{...}{optional parameters for the mean function}
}
\value{
Function returns numerical value for each column (sample) of provided data frame
}
\description{
It calulates aboundance scores through a mean of marker genes
}
\examples{

#get_scores (gene_expression, my_markers.list, summary = "mean", na.rm = TRUE)
}
