% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_test_samples.R
\name{most_variant_IC}
\alias{most_variant_IC}
\title{Compute variance explained by each Independent Component}
\usage{
most_variant_IC(S, A, X, n = 5)
}
\arguments{
\item{S}{result of \code{\link{run_fastica}} the \code{S} matrix}

\item{A}{result of \code{\link{run_fastica}} the \code{A} matrix}

\item{X}{data, either post-PCA data of \code{\link{run_fastica}} \code{X} matrix}

\item{n}{number of top ICs if \code{n} = "all" then fev for all ICs is returned}
}
\value{
returns a data frame with n top ICs numbers ranked by their fraction of variance explained
}
\description{
Compute variance explained by each Independent Component
}
\examples{


}
