% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastica.R
\name{run_fastica}
\alias{run_fastica}
\title{Decompose dataset with ICA.}
\usage{
run_fastica(X, optimal = TRUE, row.center = TRUE, with.names = FALSE,
  gene.names = NULL, alg.typ = "parallel", method = "C", n.comp = 100,
  isLog = TRUE, R = TRUE, path_global = getwd(),
  matlbpth = matlabr::get_matlab(),
  fasticapth = paste0(path.package("deconica", quiet = TRUE), "/fastica++"),
  export.corr = FALSE, ...)
}
\arguments{
\item{X}{a data matrix with \code{n} rows representing observations and
\code{p} columns representing variables, place gene names in the first
column and select \code{with.names = TRUE}}

\item{optimal}{check \code{TRUE} to let select best number of components for
deconvolution, for datasets >120 columns, n.comp will be set to 100, if <120
then number of components will be selected according to Kaiser Rule (90
percent of variance explained)}

\item{row.center}{if \code{TRUE} subtract row mean from data}

\item{with.names}{if first column of X is row.names please indicate
\code{TRUE}, in case of duplicated names, the transcript with highest
variance will be kept, names need to be HUGO names, if names are not
provided at this step, you can provide them later}

\item{gene.names}{character vector of row names - gene names}

\item{alg.typ}{if \code{alg.typ == "parallel"} the components are extracted
    simultaneously (the default). if \code{alg.typ == "deflation"} the
    components are extracted one at a time.}

\item{method}{if \code{method == "R"} then computations are done
    exclusively in \R (default). The code allows the interested \R user to
    see exactly what the algorithm does.
    if \code{method == "C"} then C code is used to perform most of the
    computations, which makes the algorithm run faster.  During
    compilation the C code is linked to an optimized BLAS library if
    present, otherwise stand-alone BLAS routines are compiled.}

\item{n.comp}{number of components to be extracted}

\item{isLog}{if data is in log \code{TRUE} if data is in counts \code{FALSE}}

\item{R}{if TRUE (default) the R version of fastICA is running, else the
matlab version (you need to provide parameters of your matlab engine)}

\item{path_global}{only if \code{R = FALSE}, the global path where files will
be written, current directory by default}

\item{matlbpth}{only if \code{R = FALSE},  the path to matlab engine, it uses
\code{\link[matlabr:get_matlab]{get_matlab}} to find path to your matlab automatically}

\item{fasticapth}{path to repository of source matlab code, it is set by default
as coming with the package}

\item{export.corr}{\code{TRUE} if you need to export \code{S} matrix in a specific
format for correlation in external java app}

\item{...}{other possible parameters for \code{\link[fastICA]{fastICA}}}
}
\value{
A list containing the following components as in
 \code{\link[fastICA]{fastICA}} \describe{
 \item{X}{pre-processed data matrix
 (after PCA)}
 \item{K}{pre-whitening matrix that projects data onto the first
 n.comp principal components.}
 \item{W}{estimated un-mixing matrix (see definition in details)}
 \item{A}{estimated mixing matrix}
 \item{S}{estimated source matrix}
  \item{names}{if  \code{with.names = TRUE} will contain row
 names list}
 \item{counts}{if  \code{isLog = FALSE} will contain initial
 matrix without duplicated genes}
 \item{log.counts}{initial matrix without
 duplicated genes in log2(x+1) before centering}
 \item{samples}{sample names as provided}
 }
}
\description{
This is a wrapper of \code{\link[fastICA]{fastICA}}. It allows compute number
of ICs optimal for over decomposition for immune deconvolution.
}
\examples{

# numerical matrix
S <- matrix(runif(10000), 5000, 2)
A <- matrix(c(1, 1, -1, 3), 2, 2, byrow = TRUE)
X <- data.frame(S \%*\% A)
run_fastica(X, row.center = TRUE, n.comp = 2, optimal = FALSE)
#matlab
\dontrun{
run_fastica(X, row.center = TRUE, n.comp = 2, optimal = FALSE, R = FALSE)
}
# matrix with gene names
S <- matrix(runif(10000), 5000, 2)
A <- matrix(c(1, 1, -1, 3), 2, 2, byrow = TRUE)
X <- data.frame(S \%*\% A)
names <- paste("A",1:nrow(X), sep="")
X <- cbind(names,X)
run_fastica(X, row.center = TRUE, n.comp = 2, optimal = FALSE, with.names = TRUE)

}
\seealso{
\code{\link[fastICA]{fastICA}}
 \url{https://cran.r-project.org/web/packages/fastICA/index.html}
}
