% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{scores_corr_plot}
\alias{scores_corr_plot}
\title{Correlation plot of abundance scores}
\usage{
scores_corr_plot(x, y, ...)
}
\arguments{
\item{x}{\code{matrix} or \code{data.frame} of abundance scores, samples in rows and cell types in columns}

\item{y}{\code{matrix} or \code{data.frame} of expected (or to compare) abundance scores, samples in rows and cell types in columns}

\item{...}{additional parameters for \code{method} from \code{\link[corrplot:corrplot]{corrplot}}}
}
\value{
\describe{\item{correlation plot}{based on \code{\link[corrplot:corrplot]{corrplot}}}
\item{\code{corr.full}}{full correlation matrix}
\item{\code{corr.filtere}}{correlation without correlation with itself}}
}
\description{
Produces correlation plot of abundance scores estimated versus expected
}
\details{
correlation plot between different abundance scores of cell types in samples, correlates both matrices with each other
merging two \code{data.frame}s by \code{row.name}s, on \code{\link[corrplot:corrplot]{corrplot}} \code{is.corr} parameter is
set to \code{FALSE}
}
\examples{
x <- matrix(runif(1000), ncol = 10, nrow = 10)
y <- matrix(runif(1000), ncol = 10, nrow = 10)
row.names(x) <- row.names(y) <- paste0("S", 1:10)
colnames(x) <- paste0("CL_",  1:10, "_estimated")
colnames(y) <- paste0("CL_",  1:10, "_expected")
scores_corr_plot(x,y, method = "number", tl.col = "black")
scores_corr_plot(x,y, method = "square", tl.col = "black")
}
\seealso{
\code{\link[Hmisc:rcorr]{rcorr}}, \code{\link[corrplot:corrplot]{corrplot}}
}
